requirements.txt
asn1crypto==0.23.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.2
dj-database-url==0.4.1
Django==1.11.1
django-bmemcached==0.2.3
django-slack-oauth==1.3.0
enum34==1.1.6
google-apputils==0.4.2
gunicorn==19.6.0
idna==2.6
ipaddress==1.0.18
jsonfield==2.0.2
protobuf==3.3.0
psycopg2==2.6.2
pycparser==2.18
python-binary-memcached==0.26.1
python-dateutil==2.6.1
python-gflags==3.1.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
whitenoise==3.2
